backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain-openai 
langchain-google-genai 
langchain-groq